added utils, dtos, converters, IncidentService, changed incident controller and repository

+++ back-end/Application/src/main/java/com/controlroom/Application/repository/IncidentRepository.java
@@ -10,7 +10,7 @@ import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface IncidentRepository extends JpaRepository<Incident, Integer> {
+public interface IncidentRepository extends JpaRepository<Incident, Long> {
 
     List<Incident> findAll();
